+++ lib/cg/Transformations.cpp
@@ -20,32 +20,32 @@ namespace CG {
 		return lhs;
 	}
 	
-	Transformation Transformation::newTranslation(float dx, float dy){
+	Transformation Transformation::newTranslation(double dx, double dy){
 		return Transformation
 			({{ {1,  0,  0},
 				{0,  1,  0},
 				{dx, dy, 1}}});
 	}
 
-	Transformation Transformation::newScaling(float sx, float sy){
+	Transformation Transformation::newScaling(double sx, double sy){
 		return Transformation
 			({{ {sx, 0 , 0},
 				{0 , sy, 0},
 				{0 , 0 , 1}}});
 	}
 
-	Transformation Transformation::newRotationAroundOrigin(float theta){
+	Transformation Transformation::newRotationAroundOrigin(double theta){
 		return Transformation
-			({{ {cosf(theta), -sinf(theta), 0},
-				{sinf(theta), cosf(theta) , 0},
+			({{ {cos(theta), -sin(theta), 0},
+				{sin(theta), cos(theta) , 0},
 				{0         , 0          , 1}}});
 	}
 
-	Transformation Transformation::newRotationAroundPoint(float theta, const Coordinate &p){
+	Transformation Transformation::newRotationAroundPoint(double theta, const Coordinate &p){
 		return newTranslation(-p.x, -p.y) * newRotationAroundOrigin(theta) * newTranslation(p.x, p.y);
 	}
 
-	Transformation Transformation::newRotationAroundCenterOfMass(float theta, const GObject& obj){
+	Transformation Transformation::newRotationAroundCenterOfMass(double theta, const GObject& obj){
 		const Coordinate center = obj.center();
 		return newRotationAroundPoint(theta, center);
 	}
