+++ lib/ui/ObjectsTreeView.cpp
@@ -107,7 +107,7 @@ void ObjectsTreeView::on_menu_popup_rotate() {
   RotateDialog dialog(objCenter);
 
   if (dialog.run() == Gtk::RESPONSE_OK) {
-    float degrees = dialog.getRotation();
+    double degrees = dialog.getRotation();
     CG::Coordinate rotationCenter = dialog.getRotationCenter();
     CG::Transformation transformation = CG::Transformation::newRotationAroundPoint(CG::Transformation::toRadians(degrees), rotationCenter);
     obj.transform(transformation);
